test: isolate HyperCore cost CLI coverage
